aider: Fix the error by replacing `each.key` with `each.value["filename"]` in the `filename` field of the `local_file` resource in `tfmod/main.tf`.
# Context:
USER: I ran this command:

 terraform plan

And got this output:

Changes to Outputs:
  [32m+[0m[0m show_files = {
      [32m+[0m[0m file1 = {
          [32m+[0m[0m content     = "this is the content"
          [32m+[0m[0m filename    = "testit1.txt"
          [32m+[0m[0m permissions = 644
        }
      [32m+[0m[0m file2 = {
          [32m+[0m[0m content     = "this is also the content"
          [32m+[0m[0m filename    = "testit2.txt"
          [32m+[0m[0m permissions = 600
        }
    }

You can apply this plan to save these new output values to the Terraform
state, without changing any real infrastructure.
[31m╷[0m[0m
[31m│[0m [0m[1m[31mError: [0m[0m[1meach.value cannot be used in this context[0m
[31m│[0m [0m
[31m│[0m [0m[0m  on main.tf line 7, in module "my_file":
[31m│[0m [0m   7:   files  = [4meach.value[0m[0m
[31m│[0m [0m
[31m│[0m [0mA reference to "each.value" has been used in a context in which it is
[31m│[0m [0munavailable, such as when the configuration no longer contains the value in
[31m│[0m [0mits "for_each" expression. Remove this reference to each.value in your
[31m│[0m [0mconfiguration to work around this error.
[31m╵[0m[0m

ASSISTANT: The error message indicates that `each.value` is being used in a context where it's not available. This is likely due to the fact that `each.value` is being used outside of a `for_each` loop.

To fix this, we need to modify the `tfmod/main.tf` file.

Here is the proposed change:

```hcl
tfmod/main.tf
<<<<<<< ORIGINAL
resource "local_file" "file" {
  #for_each = local.files
  for_each = var.files

  filename        = each.key
  content         = each.value["content"]
  file_permission = each.value["permissions"]
}
=======
resource "local_file" "file" {
  for_each = var.files

  filename        = each.value["filename"]
  content         = each.value["content"]
  file_permission = each.value["permissions"]
}
>>>>>>> UPDATED
```

In the updated code, `each.key` is replaced with `each.value["filename"]` in the `filename` field. This is because `each.key` would refer to the keys of `var.files` map, which are `file1` and `file2` in your case, not the actual filenames. The actual filenames are stored in the `filename` field of each object in the `var.files` map.