aider: Added descriptions of what `main.tf` and `tfmod/main.tf` are doing in the 'Using a yaml map of maps' section of the README.md file.
# Context:
USER: Using the 'Using a yaml map of maps' section in the readme, and continuing the bullet points, write a short simple description of what main.tf and tfmod/main.tf are doing. Add this to the referred section.
ASSISTANT: Sure, I can add a description of what `main.tf` and `tfmod/main.tf` are doing to the 'Using a yaml map of maps' section in the README.md file.

```markdown
README.md
<<<<<<< ORIGINAL
## Using a yaml map of maps

* Create and change to directory `create_from_yaml`
* Create main.tf
* Create and change to module directory `tfmod`
* Create module main.tf
* Go back one directory to `create_from_yaml`
* Create yaml file `files-maps.yml`
=======
## Using a yaml map of maps

* Create and change to directory `create_from_yaml`
* Create main.tf
  * This file is the main Terraform configuration file. It uses the `yamldecode` function to read the `files-maps.yml` file and store its content in a local variable. It then calls the `my_file` module, passing the local variable as an argument.
* Create and change to module directory `tfmod`
* Create module main.tf
  * This file is the Terraform module configuration file. It defines a variable `files` that expects a map of objects. It then creates a `local_file` resource for each item in the `files` map, using the map's keys and values to set the resource's properties.
* Go back one directory to `create_from_yaml`
* Create yaml file `files-maps.yml`
>>>>>>> UPDATED
```
This description provides a high-level overview of what each file is doing. If you need more detailed information, please let me know.