aider: Expanded on the section "Read a list of maps from YAML" in the readme.
# Context:
USER: Please expand on the section "Read a list of maps from YAML" in the readme.
ASSISTANT: Sure, I will expand on the section "Read a list of maps from YAML" in the README.md file.

Here is the change:

```markdown
README.md
<<<<<<< ORIGINAL
## Read a list of maps from YAML

!!! Use list_of_maps directory and show_first directories.

* Modify `files.yml` to have a list of maps.
* terraform init && terraform fmt && terraform validate
* add error output
* comment module call and run init, fmt, validate again
* terraform plan, show output
=======
## Read a list of maps from YAML

!!! Use list_of_maps directory and show_first directories.

In some cases, your YAML data might be structured as a list of maps. This is a common format for representing a collection of similar objects. Each map in the list represents an object, and the keys and values in the map represent the properties of the object.

Here is an example of a list of maps in YAML:

```yaml
- filename: "file1.txt"
  content: "Hello, World!"
  permissions: "0644"
- filename: "file2.txt"
  content: "Hello, Terraform!"
  permissions: "0644"
```

In this example, each map in the list represents a file. The `filename` key specifies the name of the file, the `content` key specifies the content of the file, and the `permissions` key specifies the permissions of the file.

To read a list of maps from a YAML file in Terraform, you can use the `yamldecode` function, similar to how you would read a map of maps. However, because the `yamldecode` function returns a dynamic data type, you need to ensure that your Terraform code can handle a list of maps.

* Modify `files.yml` to have a list of maps.
* terraform init && terraform fmt && terraform validate
* If there are any errors, add the error output to your code for debugging.
* Comment out the module call and run `terraform init`, `terraform fmt`, and `terraform validate` again to check for errors.
* Run `terraform plan` to see the planned changes. The output should show the list of maps that was read from the `files.yml` file.
>>>>>>> UPDATED
```