aider: Add code for "Using a yaml map of maps" section to the readme as fenced code blocks and fix any errors.
# Context:
USER: In the readme, lines beginning with `!!!` are notes to me, the developer of this tutorial, and will not be included in the finished project.
ASSISTANT: Understood. I will ignore lines beginning with `!!!` as they are developer notes and not intended for the final version of the README.md. How can I assist you further?
USER: Please use the !!! lines to inform you of how certain things should go (for example, the 'use somedirectory' is a not that the code for this section is in that directory).
ASSISTANT: Understood. I will use the lines beginning with `!!!` as guidance for understanding the structure and organization of the tutorial. How can I assist you further?
USER: Add the code referenced in the "Using a yaml map of maps" and fix any errors you see.
ASSISTANT: To add the code referenced in the "Using a yaml map of maps" section, we need to modify the following files:

1. map_of_maps/main.tf
2. tfmod/main.tf

Here are the changes:

```hcl
map_of_maps/main.tf
<<<<<<< ORIGINAL
locals {
  files = yamldecode(file("files.yml"))
}

module "my_file" {
  source = "./tfmod"
  files  = local.files
}

output "show_files" {
  value = local.files
}
=======
locals {
  files =
^C KeyboardInterrupt
USER: Add the code referenced in the "Using a yaml map of maps" section to the readme as fenced code blocks and fix any errors you see.
ASSISTANT: To add the code referenced in the "Using a yaml map of maps" section to the README.md file, we need to modify the README.md file.

Here are the changes:

```markdown
README.md
<<<<<<< ORIGINAL
## Using a yaml map of maps

!!! Use map_of_maps directory

* Create and change to directory `create_from_yaml`
* Create main.tf
* Add code
  * This file is the main Terraform configuration file. It uses the
    `yamldecode` function to read the `files.yml` file and store its
    content in a local variable. It then calls the `my_file` module, passing
    the local variable as an argument.

* Create and change to module directory `tfmod`
* Create module main.tf
* Add code
  * This file is the Terraform module configuration file. It defines
    a variable `files` that expects a map of objects. It then creates
    a `local_file` resource for each item in the `files` map, using the map's
    keys and values to set the resource's properties.

* Go back one directory to `create_from_yaml`
* Create yaml file `files.yml`
* terraform init && terraform fmt && terraform validate
* terraform plan
=======
## Using a yaml map of maps

!!! Use map_of_maps directory

* Create and change to directory `create_from_yaml`
* Create main.tf
* Add code
  * This file is the main Terraform configuration file. It uses the
    `yamldecode` function to read the `files.yml` file and store its
    content in a local variable. It then calls the `my_file` module, passing
    the local variable as an argument.

```hcl
locals {
  files = yamldecode(file("files.yml"))
}

module "my_file" {
  source = "./tfmod"
  files  = local.files
}

output "show_files" {
  value = local.files
}
```

* Create and change to module directory `tfmod`
* Create module main.tf
* Add code
  * This file is the Terraform module configuration file. It defines
    a variable `files` that expects a map of objects. It then creates
    a `local_file` resource for each item in the `files` map, using the map's
    keys and values to set the resource's properties.

```hcl
terraform {
  required_providers {
    local = { source = "hashicorp/local" }
  }
}

variable "files" {
  type = map(object({
    filename    = string,
    content     = string,
    permissions = string
  }))
}

locals {
  files = { for file in var.files : file["filename"] => file }
}

resource "local_file" "file" {
  for_each = local.files

  filename        = each.key
  content         = each.value["content"]
  file_permission = each.value["permissions"]
}
```

* Go back one directory to `create_from_yaml`
* Create yaml file `files.yml`
* terraform init && terraform fmt && terraform validate
* terraform plan
>>>>>>> UPDATED
```