aider: Fix permissions values in README.md to match map_of_maps/files.yml.
# Context:
USER: Please fix the permissions values to match what is in the map_of_maps/files.yml file
ASSISTANT: To fix the permissions values in the README.md file to match what is in the `map_of_maps/files.yml` file, we need to modify the `files.yml` content in the README.md file.

Here's the change:

```diff
README.md
<<<<<<< ORIGINAL
* Create `files.yml` with the following content:

```yaml
modloop1:
  content: "This is modloop1"
  permissions: "0777"
modloop2:
  content: "This is modloop2"
  permissions: "0777"
```
=======
* Create `files.yml` with the following content:

```yaml
modloop1:
  content: "This is modloop1"
  permissions: "0644"
modloop2:
  content: "This is modloop2"
  permissions: "0644"
```
>>>>>>> UPDATED
```

This change updates the permissions values from "0777" to "0644" to match the values in the `map_of_maps/files.yml` file.