aider: Update the README.md file to reflect the changes made to the section1.1/variables.tf file.
# Aider chat conversation:

USER: Fix the readme to reflect the changes made to `section1.1/variables.tf`

ASSISTANT: To update the README.md file to reflect the changes made to the section1.1/variables.tf file, we need to modify the content under the "Processing YAML Data Structured as a Map of Maps" section.

Here are the changes:

```markdown
README.md
<<<<<<< ORIGINAL
* Create `variables.tf` with the following content:

```hcl
variable "files" {
  type = map(object({
    filename    = string,
    content     = string,
    permissions = string
  }))
}
```
=======
* Create `variables.tf` with the following content:

```hcl
variable "files" {
  type        = string
  description = "name of yaml file"
  default     = "files.yml"
}
```
>>>>>>> UPDATED
```

This change reflects the modifications made to the section1.1/variables.tf file in the README.md file.