aider: Add links to the fileset, flatten, and merge function documentation to the "Additional Information" section of the readme.
# Aider chat conversation:

USER: Add links to the fileset, flatten and merge function documentation to the "Additional Information section" of the readme.

ASSISTANT: Sure, I can add links to the Terraform documentation for the `fileset`, `flatten`, and `merge` functions in the "Additional Information" section of the README.md file. Here's the proposed change:

```markdown
README.md
<<<<<<< ORIGINAL
## Additional Information
=======
## Additional Information

For more information on the Terraform functions used in this tutorial, you can refer to the official Terraform documentation:

- [`fileset`](https://www.terraform.io/docs/language/functions/fileset.html): Generates a list of files that match a given pattern in a given directory.
- [`flatten`](https://www.terraform.io/docs/language/functions/flatten.html): Takes a list and replaces any elements that are lists with a flattened sequence of the list contents.
- [`merge`](https://www.terraform.io/docs/language/functions/merge.html): Takes an arbitrary number of maps or objects, and returns a single map or object that includes a combination of all the elements of the given maps or objects.
>>>>>>> UPDATED
```